+++ module/Core/src/Paginator/Adapter/OrphanVisitsPaginatorAdapter.php
@@ -6,6 +6,8 @@ namespace Shlinkio\Shlink\Core\Paginator\Adapter;
 
 use Shlinkio\Shlink\Core\Model\VisitsParams;
 use Shlinkio\Shlink\Core\Repository\VisitRepositoryInterface;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
 
 class OrphanVisitsPaginatorAdapter extends AbstractCacheableCountPaginatorAdapter
 {
@@ -20,11 +22,20 @@ class OrphanVisitsPaginatorAdapter extends AbstractCacheableCountPaginatorAdapte
 
     protected function doCount(): int
     {
-        return $this->repo->countOrphanVisits($this->params->getDateRange());
+        return $this->repo->countOrphanVisits(new VisitsCountFiltering(
+            $this->params->getDateRange(),
+            $this->params->excludeBots(),
+        ));
     }
 
     public function getSlice($offset, $length): iterable // phpcs:ignore
     {
-        return $this->repo->findOrphanVisits($this->params->getDateRange(), $length, $offset);
+        return $this->repo->findOrphanVisits(new VisitsListFiltering(
+            $this->params->getDateRange(),
+            $this->params->excludeBots(),
+            null,
+            $length,
+            $offset,
+        ));
     }
 }