+++ module/Core/src/Paginator/Adapter/VisitsPaginatorAdapter.php
@@ -8,6 +8,8 @@ use Happyr\DoctrineSpecification\Specification\Specification;
 use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\Model\VisitsParams;
 use Shlinkio\Shlink\Core\Repository\VisitRepositoryInterface;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
 
 class VisitsPaginatorAdapter extends AbstractCacheableCountPaginatorAdapter
 {
@@ -33,10 +35,13 @@ class VisitsPaginatorAdapter extends AbstractCacheableCountPaginatorAdapter
         return $this->visitRepository->findVisitsByShortCode(
             $this->identifier->shortCode(),
             $this->identifier->domain(),
-            $this->params->getDateRange(),
-            $length,
-            $offset,
-            $this->spec,
+            new VisitsListFiltering(
+                $this->params->getDateRange(),
+                $this->params->excludeBots(),
+                $this->spec,
+                $length,
+                $offset,
+            ),
         );
     }
 
@@ -45,8 +50,11 @@ class VisitsPaginatorAdapter extends AbstractCacheableCountPaginatorAdapter
         return $this->visitRepository->countVisitsByShortCode(
             $this->identifier->shortCode(),
             $this->identifier->domain(),
-            $this->params->getDateRange(),
-            $this->spec,
+            new VisitsCountFiltering(
+                $this->params->getDateRange(),
+                $this->params->excludeBots(),
+                $this->spec,
+            ),
         );
     }
 }