+++ module/Core/src/Repository/VisitRepositoryInterface.php
@@ -6,9 +6,9 @@ namespace Shlinkio\Shlink\Core\Repository;
 
 use Doctrine\Persistence\ObjectRepository;
 use Happyr\DoctrineSpecification\Repository\EntitySpecificationRepositoryInterface;
-use Happyr\DoctrineSpecification\Specification\Specification;
-use Shlinkio\Shlink\Common\Util\DateRange;
 use Shlinkio\Shlink\Core\Entity\Visit;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
 use Shlinkio\Shlink\Rest\Entity\ApiKey;
 
 interface VisitRepositoryInterface extends ObjectRepository, EntitySpecificationRepositoryInterface
@@ -33,41 +33,23 @@ interface VisitRepositoryInterface extends ObjectRepository, EntitySpecification
     /**
      * @return Visit[]
      */
-    public function findVisitsByShortCode(
-        string $shortCode,
-        ?string $domain = null,
-        ?DateRange $dateRange = null,
-        ?int $limit = null,
-        ?int $offset = null,
-        ?Specification $spec = null
-    ): array;
+    public function findVisitsByShortCode(string $shortCode, ?string $domain, VisitsListFiltering $filtering): array;
 
-    public function countVisitsByShortCode(
-        string $shortCode,
-        ?string $domain = null,
-        ?DateRange $dateRange = null,
-        ?Specification $spec = null
-    ): int;
+    public function countVisitsByShortCode(string $shortCode, ?string $domain, VisitsCountFiltering $filtering): int;
 
     /**
      * @return Visit[]
      */
-    public function findVisitsByTag(
-        string $tag,
-        ?DateRange $dateRange = null,
-        ?int $limit = null,
-        ?int $offset = null,
-        ?Specification $spec = null
-    ): array;
+    public function findVisitsByTag(string $tag, VisitsListFiltering $filtering): array;
 
-    public function countVisitsByTag(string $tag, ?DateRange $dateRange = null, ?Specification $spec = null): int;
+    public function countVisitsByTag(string $tag, VisitsCountFiltering $filtering): int;
 
     /**
      * @return Visit[]
      */
-    public function findOrphanVisits(?DateRange $dateRange = null, ?int $limit = null, ?int $offset = null): array;
+    public function findOrphanVisits(VisitsListFiltering $filtering): array;
 
-    public function countOrphanVisits(?DateRange $dateRange = null): int;
+    public function countOrphanVisits(VisitsCountFiltering $filtering): int;
 
     public function countVisits(?ApiKey $apiKey = null): int;
 }