+++ module/Core/test-db/Repository/VisitRepositoryTest.php
@@ -15,6 +15,8 @@ use Shlinkio\Shlink\Core\Model\ShortUrlMeta;
 use Shlinkio\Shlink\Core\Model\Visitor;
 use Shlinkio\Shlink\Core\Repository\VisitRepository;
 use Shlinkio\Shlink\Core\ShortUrl\Resolver\PersistenceShortUrlRelationResolver;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
 use Shlinkio\Shlink\IpGeolocation\Model\Location;
 use Shlinkio\Shlink\Rest\ApiKey\Model\ApiKeyMeta;
 use Shlinkio\Shlink\Rest\ApiKey\Model\RoleDefinition;
@@ -87,22 +89,34 @@ class VisitRepositoryTest extends DatabaseTestCase
     {
         [$shortCode, $domain] = $this->createShortUrlsAndVisits();
 
-        self::assertCount(0, $this->repo->findVisitsByShortCode('invalid'));
-        self::assertCount(6, $this->repo->findVisitsByShortCode($shortCode));
-        self::assertCount(3, $this->repo->findVisitsByShortCode($shortCode, $domain));
-        self::assertCount(2, $this->repo->findVisitsByShortCode($shortCode, null, new DateRange(
-            Chronos::parse('2016-01-02'),
-            Chronos::parse('2016-01-03'),
+        self::assertCount(0, $this->repo->findVisitsByShortCode('invalid', null, new VisitsListFiltering()));
+        self::assertCount(6, $this->repo->findVisitsByShortCode($shortCode, null, new VisitsListFiltering()));
+        self::assertCount(4, $this->repo->findVisitsByShortCode($shortCode, null, new VisitsListFiltering(null, true)));
+        self::assertCount(3, $this->repo->findVisitsByShortCode($shortCode, $domain, new VisitsListFiltering()));
+        self::assertCount(2, $this->repo->findVisitsByShortCode($shortCode, null, new VisitsListFiltering(
+            DateRange::withStartAndEndDate(Chronos::parse('2016-01-02'), Chronos::parse('2016-01-03')),
         )));
-        self::assertCount(4, $this->repo->findVisitsByShortCode($shortCode, null, new DateRange(
-            Chronos::parse('2016-01-03'),
+        self::assertCount(4, $this->repo->findVisitsByShortCode($shortCode, null, new VisitsListFiltering(
+            DateRange::withStartDate(Chronos::parse('2016-01-03')),
         )));
-        self::assertCount(1, $this->repo->findVisitsByShortCode($shortCode, $domain, new DateRange(
-            Chronos::parse('2016-01-03'),
+        self::assertCount(1, $this->repo->findVisitsByShortCode($shortCode, $domain, new VisitsListFiltering(
+            DateRange::withStartDate(Chronos::parse('2016-01-03')),
         )));
-        self::assertCount(3, $this->repo->findVisitsByShortCode($shortCode, null, null, 3, 2));
-        self::assertCount(2, $this->repo->findVisitsByShortCode($shortCode, null, null, 5, 4));
-        self::assertCount(1, $this->repo->findVisitsByShortCode($shortCode, $domain, null, 3, 2));
+        self::assertCount(3, $this->repo->findVisitsByShortCode(
+            $shortCode,
+            null,
+            new VisitsListFiltering(null, false, null, 3, 2),
+        ));
+        self::assertCount(2, $this->repo->findVisitsByShortCode(
+            $shortCode,
+            null,
+            new VisitsListFiltering(null, false, null, 5, 4),
+        ));
+        self::assertCount(1, $this->repo->findVisitsByShortCode(
+            $shortCode,
+            $domain,
+            new VisitsListFiltering(null, false, null, 3, 2),
+        ));
     }
 
     /** @test */
@@ -110,18 +124,21 @@ class VisitRepositoryTest extends DatabaseTestCase
     {
         [$shortCode, $domain] = $this->createShortUrlsAndVisits();
 
-        self::assertEquals(0, $this->repo->countVisitsByShortCode('invalid'));
-        self::assertEquals(6, $this->repo->countVisitsByShortCode($shortCode));
-        self::assertEquals(3, $this->repo->countVisitsByShortCode($shortCode, $domain));
-        self::assertEquals(2, $this->repo->countVisitsByShortCode($shortCode, null, new DateRange(
-            Chronos::parse('2016-01-02'),
-            Chronos::parse('2016-01-03'),
+        self::assertEquals(0, $this->repo->countVisitsByShortCode('invalid', null, new VisitsCountFiltering()));
+        self::assertEquals(6, $this->repo->countVisitsByShortCode($shortCode, null, new VisitsCountFiltering()));
+        self::assertEquals(4, $this->repo->countVisitsByShortCode($shortCode, null, new VisitsCountFiltering(
+            null,
+            true,
         )));
-        self::assertEquals(4, $this->repo->countVisitsByShortCode($shortCode, null, new DateRange(
-            Chronos::parse('2016-01-03'),
+        self::assertEquals(3, $this->repo->countVisitsByShortCode($shortCode, $domain, new VisitsCountFiltering()));
+        self::assertEquals(2, $this->repo->countVisitsByShortCode($shortCode, null, new VisitsCountFiltering(
+            DateRange::withStartAndEndDate(Chronos::parse('2016-01-02'), Chronos::parse('2016-01-03')),
         )));
-        self::assertEquals(1, $this->repo->countVisitsByShortCode($shortCode, $domain, new DateRange(
-            Chronos::parse('2016-01-03'),
+        self::assertEquals(4, $this->repo->countVisitsByShortCode($shortCode, null, new VisitsCountFiltering(
+            DateRange::withStartDate(Chronos::parse('2016-01-03')),
+        )));
+        self::assertEquals(1, $this->repo->countVisitsByShortCode($shortCode, $domain, new VisitsCountFiltering(
+            DateRange::withStartDate(Chronos::parse('2016-01-03')),
         )));
     }
 
@@ -140,13 +157,15 @@ class VisitRepositoryTest extends DatabaseTestCase
         $this->createShortUrlsAndVisits(false, [$foo]);
         $this->getEntityManager()->flush();
 
-        self::assertCount(0, $this->repo->findVisitsByTag('invalid'));
-        self::assertCount(18, $this->repo->findVisitsByTag($foo));
-        self::assertCount(6, $this->repo->findVisitsByTag($foo, new DateRange(
-            Chronos::parse('2016-01-02'),
-            Chronos::parse('2016-01-03'),
+        self::assertCount(0, $this->repo->findVisitsByTag('invalid', new VisitsListFiltering()));
+        self::assertCount(18, $this->repo->findVisitsByTag($foo, new VisitsListFiltering()));
+        self::assertCount(12, $this->repo->findVisitsByTag($foo, new VisitsListFiltering(null, true)));
+        self::assertCount(6, $this->repo->findVisitsByTag($foo, new VisitsListFiltering(
+            DateRange::withStartAndEndDate(Chronos::parse('2016-01-02'), Chronos::parse('2016-01-03')),
         )));
-        self::assertCount(12, $this->repo->findVisitsByTag($foo, new DateRange(Chronos::parse('2016-01-03'))));
+        self::assertCount(12, $this->repo->findVisitsByTag($foo, new VisitsListFiltering(
+            DateRange::withStartDate(Chronos::parse('2016-01-03')),
+        )));
     }
 
     /** @test */
@@ -160,13 +179,15 @@ class VisitRepositoryTest extends DatabaseTestCase
         $this->createShortUrlsAndVisits(false, [$foo]);
         $this->getEntityManager()->flush();
 
-        self::assertEquals(0, $this->repo->countVisitsByTag('invalid'));
-        self::assertEquals(12, $this->repo->countVisitsByTag($foo));
-        self::assertEquals(4, $this->repo->countVisitsByTag($foo, new DateRange(
-            Chronos::parse('2016-01-02'),
-            Chronos::parse('2016-01-03'),
+        self::assertEquals(0, $this->repo->countVisitsByTag('invalid', new VisitsCountFiltering()));
+        self::assertEquals(12, $this->repo->countVisitsByTag($foo, new VisitsCountFiltering()));
+        self::assertEquals(8, $this->repo->countVisitsByTag($foo, new VisitsCountFiltering(null, true)));
+        self::assertEquals(4, $this->repo->countVisitsByTag($foo, new VisitsCountFiltering(
+            DateRange::withStartAndEndDate(Chronos::parse('2016-01-02'), Chronos::parse('2016-01-03')),
         )));
-        self::assertEquals(8, $this->repo->countVisitsByTag($foo, new DateRange(Chronos::parse('2016-01-03'))));
+        self::assertEquals(8, $this->repo->countVisitsByTag($foo, new VisitsCountFiltering(
+            DateRange::withStartDate(Chronos::parse('2016-01-03')),
+        )));
     }
 
     /** @test */
@@ -206,6 +227,7 @@ class VisitRepositoryTest extends DatabaseTestCase
         $this->getEntityManager()->persist(Visit::forBasePath(Visitor::emptyInstance()));
         $this->getEntityManager()->persist(Visit::forInvalidShortUrl(Visitor::emptyInstance()));
         $this->getEntityManager()->persist(Visit::forRegularNotFound(Visitor::emptyInstance()));
+        $this->getEntityManager()->persist(Visit::forRegularNotFound(Visitor::botInstance()));
 
         $this->getEntityManager()->flush();
 
@@ -213,7 +235,8 @@ class VisitRepositoryTest extends DatabaseTestCase
         self::assertEquals(4, $this->repo->countVisits($apiKey1));
         self::assertEquals(5 + 7, $this->repo->countVisits($apiKey2));
         self::assertEquals(4 + 7, $this->repo->countVisits($domainApiKey));
-        self::assertEquals(3, $this->repo->countOrphanVisits());
+        self::assertEquals(4, $this->repo->countOrphanVisits(new VisitsCountFiltering()));
+        self::assertEquals(3, $this->repo->countOrphanVisits(new VisitsCountFiltering(null, true)));
     }
 
     /** @test */
@@ -223,9 +246,10 @@ class VisitRepositoryTest extends DatabaseTestCase
         $this->getEntityManager()->persist($shortUrl);
         $this->createVisitsForShortUrl($shortUrl, 7);
 
+        $botsCount = 3;
         for ($i = 0; $i < 6; $i++) {
             $this->getEntityManager()->persist($this->setDateOnVisit(
-                Visit::forBasePath(Visitor::emptyInstance()),
+                Visit::forBasePath($botsCount < 1 ? Visitor::emptyInstance() : Visitor::botInstance()),
                 Chronos::parse(sprintf('2020-01-0%s', $i + 1)),
             ));
             $this->getEntityManager()->persist($this->setDateOnVisit(
@@ -236,20 +260,32 @@ class VisitRepositoryTest extends DatabaseTestCase
                 Visit::forRegularNotFound(Visitor::emptyInstance()),
                 Chronos::parse(sprintf('2020-01-0%s', $i + 1)),
             ));
+
+            $botsCount--;
         }
 
         $this->getEntityManager()->flush();
 
-        self::assertCount(18, $this->repo->findOrphanVisits());
-        self::assertCount(5, $this->repo->findOrphanVisits(null, 5));
-        self::assertCount(10, $this->repo->findOrphanVisits(null, 15, 8));
-        self::assertCount(9, $this->repo->findOrphanVisits(DateRange::withStartDate(Chronos::parse('2020-01-04')), 15));
-        self::assertCount(2, $this->repo->findOrphanVisits(
+        self::assertCount(18, $this->repo->findOrphanVisits(new VisitsListFiltering()));
+        self::assertCount(15, $this->repo->findOrphanVisits(new VisitsListFiltering(null, true)));
+        self::assertCount(5, $this->repo->findOrphanVisits(new VisitsListFiltering(null, false, null, 5)));
+        self::assertCount(10, $this->repo->findOrphanVisits(new VisitsListFiltering(null, false, null, 15, 8)));
+        self::assertCount(9, $this->repo->findOrphanVisits(new VisitsListFiltering(
+            DateRange::withStartDate(Chronos::parse('2020-01-04')),
+            false,
+            null,
+            15,
+        )));
+        self::assertCount(2, $this->repo->findOrphanVisits(new VisitsListFiltering(
             DateRange::withStartAndEndDate(Chronos::parse('2020-01-02'), Chronos::parse('2020-01-03')),
+            false,
+            null,
             6,
             4,
-        ));
-        self::assertCount(3, $this->repo->findOrphanVisits(DateRange::withEndDate(Chronos::parse('2020-01-01'))));
+        )));
+        self::assertCount(3, $this->repo->findOrphanVisits(new VisitsListFiltering(
+            DateRange::withEndDate(Chronos::parse('2020-01-01')),
+        )));
     }
 
     /** @test */
@@ -276,13 +312,17 @@ class VisitRepositoryTest extends DatabaseTestCase
 
         $this->getEntityManager()->flush();
 
-        self::assertEquals(18, $this->repo->countOrphanVisits());
-        self::assertEquals(18, $this->repo->countOrphanVisits(DateRange::emptyInstance()));
-        self::assertEquals(9, $this->repo->countOrphanVisits(DateRange::withStartDate(Chronos::parse('2020-01-04'))));
-        self::assertEquals(6, $this->repo->countOrphanVisits(
-            DateRange::withStartAndEndDate(Chronos::parse('2020-01-02'), Chronos::parse('2020-01-03')),
+        self::assertEquals(18, $this->repo->countOrphanVisits(new VisitsCountFiltering()));
+        self::assertEquals(18, $this->repo->countOrphanVisits(new VisitsCountFiltering(DateRange::emptyInstance())));
+        self::assertEquals(9, $this->repo->countOrphanVisits(
+            new VisitsCountFiltering(DateRange::withStartDate(Chronos::parse('2020-01-04'))),
         ));
-        self::assertEquals(3, $this->repo->countOrphanVisits(DateRange::withEndDate(Chronos::parse('2020-01-01'))));
+        self::assertEquals(6, $this->repo->countOrphanVisits(new VisitsCountFiltering(
+            DateRange::withStartAndEndDate(Chronos::parse('2020-01-02'), Chronos::parse('2020-01-03')),
+        )));
+        self::assertEquals(3, $this->repo->countOrphanVisits(
+            new VisitsCountFiltering(DateRange::withEndDate(Chronos::parse('2020-01-01'))),
+        ));
     }
 
     private function createShortUrlsAndVisits(bool $withDomain = true, array $tags = []): array
@@ -311,13 +351,17 @@ class VisitRepositoryTest extends DatabaseTestCase
         return [$shortCode, $domain, $shortUrl];
     }
 
-    private function createVisitsForShortUrl(ShortUrl $shortUrl, int $amount = 6): void
+    private function createVisitsForShortUrl(ShortUrl $shortUrl, int $amount = 6, int $botsAmount = 2): void
     {
         for ($i = 0; $i < $amount; $i++) {
             $visit = $this->setDateOnVisit(
-                Visit::forValidShortUrl($shortUrl, Visitor::emptyInstance()),
+                Visit::forValidShortUrl(
+                    $shortUrl,
+                    $botsAmount < 1 ? Visitor::emptyInstance() : Visitor::botInstance(),
+                ),
                 Chronos::parse(sprintf('2016-01-0%s', $i + 1)),
             );
+            $botsAmount--;
 
             $this->getEntityManager()->persist($visit);
         }