+++ module/Core/test/Mercure/MercureUpdatesGeneratorTest.php
@@ -66,6 +66,7 @@ class MercureUpdatesGeneratorTest extends TestCase
                 'userAgent' => '',
                 'visitLocation' => null,
                 'date' => $visit->getDate()->toAtomString(),
+                'potentialBot' => false,
             ],
         ], json_decode($update->getData()));
     }
@@ -91,6 +92,7 @@ class MercureUpdatesGeneratorTest extends TestCase
                 'userAgent' => '',
                 'visitLocation' => null,
                 'date' => $orphanVisit->getDate()->toAtomString(),
+                'potentialBot' => false,
                 'visitedUrl' => $orphanVisit->visitedUrl(),
                 'type' => $orphanVisit->type(),
             ],