+++ module/Core/test/Paginator/Adapter/VisitsPaginatorAdapterTest.php
@@ -12,6 +12,8 @@ use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\Model\VisitsParams;
 use Shlinkio\Shlink\Core\Paginator\Adapter\VisitsPaginatorAdapter;
 use Shlinkio\Shlink\Core\Repository\VisitRepositoryInterface;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
+use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
 use Shlinkio\Shlink\Rest\Entity\ApiKey;
 
 class VisitsPaginatorAdapterTest extends TestCase
@@ -32,9 +34,11 @@ class VisitsPaginatorAdapterTest extends TestCase
         $limit = 1;
         $offset = 5;
         $adapter = $this->createAdapter(null);
-        $findVisits = $this->repo->findVisitsByShortCode('', null, new DateRange(), $limit, $offset, null)->willReturn(
-            [],
-        );
+        $findVisits = $this->repo->findVisitsByShortCode(
+            '',
+            null,
+            new VisitsListFiltering(new DateRange(), false, null, $limit, $offset),
+        )->willReturn([]);
 
         for ($i = 0; $i < $count; $i++) {
             $adapter->getSlice($offset, $limit);
@@ -49,7 +53,11 @@ class VisitsPaginatorAdapterTest extends TestCase
         $count = 3;
         $apiKey = ApiKey::create();
         $adapter = $this->createAdapter($apiKey);
-        $countVisits = $this->repo->countVisitsByShortCode('', null, new DateRange(), $apiKey->spec())->willReturn(3);
+        $countVisits = $this->repo->countVisitsByShortCode(
+            '',
+            null,
+            new VisitsCountFiltering(new DateRange(), false, $apiKey->spec()),
+        )->willReturn(3);
 
         for ($i = 0; $i < $count; $i++) {
             $adapter->getNbResults();