+++ module/Rest/test-api/Action/TagVisitsTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace ShlinkioApiTest\Shlink\Rest\Action;
 
+use GuzzleHttp\RequestOptions;
 use Shlinkio\Shlink\TestUtils\ApiTest\ApiTestCase;
 
 use function sprintf;
@@ -14,9 +15,15 @@ class TagVisitsTest extends ApiTestCase
      * @test
      * @dataProvider provideTags
      */
-    public function expectedVisitsAreReturned(string $apiKey, string $tag, int $expectedVisitsAmount): void
-    {
-        $resp = $this->callApiWithKey(self::METHOD_GET, sprintf('/tags/%s/visits', $tag), [], $apiKey);
+    public function expectedVisitsAreReturned(
+        string $apiKey,
+        string $tag,
+        bool $excludeBots,
+        int $expectedVisitsAmount
+    ): void {
+        $resp = $this->callApiWithKey(self::METHOD_GET, sprintf('/tags/%s/visits', $tag), [
+            RequestOptions::QUERY => $excludeBots ? ['excludeBots' => true] : [],
+        ], $apiKey);
         $payload = $this->getJsonResponsePayload($resp);
 
         self::assertEquals(self::STATUS_OK, $resp->getStatusCode());
@@ -27,12 +34,16 @@ class TagVisitsTest extends ApiTestCase
 
     public function provideTags(): iterable
     {
-        yield 'foo with admin API key' => ['valid_api_key', 'foo', 5];
-        yield 'bar with admin API key' => ['valid_api_key', 'bar', 2];
-        yield 'baz with admin API key' => ['valid_api_key', 'baz', 0];
-        yield 'foo with author API key' => ['author_api_key', 'foo', 5];
-        yield 'bar with author API key' => ['author_api_key', 'bar', 2];
-        yield 'foo with domain API key' => ['domain_api_key', 'foo', 0];
+        yield 'foo with admin API key' => ['valid_api_key', 'foo', false, 5];
+        yield 'foo with admin API key and no bots' => ['valid_api_key', 'foo', true, 4];
+        yield 'bar with admin API key' => ['valid_api_key', 'bar', false, 2];
+        yield 'bar with admin API key and no bots' => ['valid_api_key', 'bar', true, 1];
+        yield 'baz with admin API key' => ['valid_api_key', 'baz', false, 0];
+        yield 'foo with author API key' => ['author_api_key', 'foo', false, 5];
+        yield 'foo with author API key and no bots' => ['author_api_key', 'foo', true, 4];
+        yield 'bar with author API key' => ['author_api_key', 'bar', false, 2];
+        yield 'bar with author API key and no bots' => ['author_api_key', 'bar', true, 1];
+        yield 'foo with domain API key' => ['domain_api_key', 'foo', false, 0];
     }
 
     /**